Fix bug in TextPickerCellRenderer on iOS
When updating items of a re-used TextPickerCellRenderer the underlying
iOS UIPickerView might access the view model based on stale view data.
This can cause index out of range errors and other inconsistencies
when the underlying model data was already updated.

The fix forces the iOS UIPickerView to reload data from the model
when it was changed before doing anything else.