Dismiss picker dialogs on dispose to fix #144
When SettingsView is disposed while a picker dialog is active we need
to explicitly dismiss it so that it is now longer shown to the user.
This does not happen automatically by disposing the dialog ref itself.
If we don't dismiss the dialog the app will crash once the user interacts
with the it, e.g. due to an ObjectDisposedException.